apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
sse-starlette==1.6.5
pydantic==2.7.0